requirements_test.txt
-r requirements.txt
pytest-homeassistant-custom-component
freezegun
